backend/app/api/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
google-generativeai
elevenlabs
python-multipart
